+++ src/assets/scripts/airport/StandardRoute/StandardRouteCollection.js
@@ -18,16 +18,17 @@ export default class StandardRouteCollection extends BaseCollection {
     /**
      * @constructor
      * @param standardRouteEnum {object}
+     * @param fixCollection {FixCollection}
      */
     /* istanbul ignore next */
-    constructor(standardRouteEnum) {
-        super();
+    constructor(standardRouteEnum, fixCollection) {
+        super(standardRouteEnum);
 
         if (typeof standardRouteEnum === 'undefined') {
             return;
         }
 
-        return this._init(standardRouteEnum);
+        return this._init(standardRouteEnum, fixCollection);
     }
 
     // TODO: refactor into a reusable class that can be fed an `item` and will be consumed by the `CanvasController`
@@ -58,10 +59,11 @@ export default class StandardRouteCollection extends BaseCollection {
      * @for StandardRouteCollection
      * @method _init
      * @param standardRouteEnum {object}
+     * @param fixCollection {FixCollection}
      * @private
      */
-    _init(standardRouteEnum) {
-        this._addRouteListToCollection(standardRouteEnum);
+    _init(standardRouteEnum, fixCollection) {
+        this._addRouteListToCollection(standardRouteEnum, fixCollection);
 
         return this;
     }
@@ -185,11 +187,12 @@ export default class StandardRouteCollection extends BaseCollection {
      * @for StandardRouteCollection
      * @method _addRouteListToCollection
      * @param routeList {object}
+     * @param fixCollection {FixCollection}
      * @private
      */
-    _addRouteListToCollection(routeList) {
+    _addRouteListToCollection(routeList, fixCollection) {
         _forEach(routeList, (route) => {
-            const routeModel = new StandardRouteModel(route);
+            const routeModel = new StandardRouteModel(route, fixCollection);
 
             this._addSidToCollection(routeModel);
         });